refactor : move wallet validation method into main internalize action validation
